Build: add -Wfloat-conversion to extra warning

+++ modules/afsocket/afsocket-source.c
@@ -41,7 +41,7 @@ int allow_severity = 0;
 int deny_severity = 0;
 #endif
 
-static const gfloat DYNAMIC_WINDOW_TIMER_MSECS = 1000;
+static const glong DYNAMIC_WINDOW_TIMER_MSECS = 1000;
 static const gsize DYNAMIC_WINDOW_REALLOC_TICKS = 5;
 
 typedef struct _AFSocketSourceConnection
@@ -328,11 +328,11 @@ afsocket_sd_set_dynamic_window_size(LogDriver *s, gint dynamic_window_size)
 }
 
 void
-afsocket_sd_set_dynamic_window_stats_freq(LogDriver *s, gfloat stats_freq)
+afsocket_sd_set_dynamic_window_stats_freq(LogDriver *s, gdouble stats_freq)
 {
   AFSocketSourceDriver *self = (AFSocketSourceDriver *) s;
 
-  self->dynamic_window_stats_freq = stats_freq * 1000.0f;
+  self->dynamic_window_stats_freq = (glong) (stats_freq * 1000);
 }
 
 void
